Extend context for tracing
